Move taskStatus.js into a modules folder for better modularity
